+++ feeluown/app.py
@@ -58,7 +58,9 @@ class App(FFrame):
         self.initialize()
 
     def initialize(self):
+        logger.debug('App start initializing...')
         self.initialized.emit()
+        logger.debug('App start initializing...done')
 
     def scan_fuo_files(self):
         fuo_files = config.FUO_FILES
@@ -91,7 +93,8 @@ class App(FFrame):
         self.player.state_changed.connect(self._on_player_status_changed)
         self.player.position_changed.connect(self._on_player_position_changed)
         self.player.duration_changed.connect(self._on_player_duration_changed)
-        self.player.media_changed.connect(self._on_player_song_changed)
+        # FIXME:
+        self.player.media_changed.connect(self._on_player_media_changed)
         self.player.playlist.playback_mode_changed.connect(
             pms_btn.on_playback_mode_changed)
 
@@ -129,7 +132,7 @@ class App(FFrame):
             painter.fillRect(self.rect(), bg_color)
 
     def _init_managers(self):
-        # self.plugins_manager.scan()
+        self.plugins_manager.scan()
         app_event_loop = asyncio.get_event_loop()
         app_event_loop.call_later(
             8, partial(asyncio.Task, self.version_manager.check_release()))
@@ -171,9 +174,12 @@ class App(FFrame):
         #     QApplication.setWindowIcon(QIcon(self.player_pixmap))
         # self.update()
 
-    def _on_player_song_changed(self, song):
+    def _on_player_media_changed(self, media):
+        # FIXME
+        return
+        song = self.player.playlist.current_song
         song_label = self.ui.top_panel.pc_panel.song_title_label
-        song_label.set_song(song.title + ' - ' + song.artists_name)
+        song_label.set_song(song.title + ' - ' + 'tmp')
 
     def _on_player_status_changed(self, state):
         pp_btn = self.ui.top_panel.pc_panel.pp_btn